Property info -Dhiraj

+++ HomeLoanApplication/src/main/java/com/sdl/homeloan/models/LocalAddress.java
@@ -22,7 +22,7 @@ public class LocalAddress {
 	private String cityName;
 	private String district;
 	private String state;
-	private double pincode;
+	private int pincode;
 	private int houseno;
 	private String streetName;
